refactor: :art: Rewrote code to adhere to coding standard
