+++ web/components/LandingPage/LPToast.tsx
@@ -1,13 +1,7 @@
-import * as React from "react";
+import React from "react";
+import { ILPToastProps } from "../../types/LPinterface";
 
-export interface ILPToastProps {
-  justifySelfEnd?: boolean;
-  textContent: string;
-}
-
-export default function LPToast(props: ILPToastProps) {
-  // TODO: Change textContent to child ?
-  const { justifySelfEnd, textContent } = props;
+const LPToast: React.FC<ILPToastProps> = ({ justifySelfEnd, textContent }) => {
   return justifySelfEnd ? (
     <div className="alert alert-success w-1/2 md:w-2/5 justify-self-end bg-apple-green">
       <div>
@@ -21,4 +15,5 @@ export default function LPToast(props: ILPToastProps) {
       </div>
     </div>
   );
-}
+};
+export default LPToast;
