+++ web/types/LPinterface.ts
@@ -1,0 +1,14 @@
+export interface ILPBodyProps { }
+export interface ILPFooterProps { }
+
+export interface ILPHeroProps {
+  bgVideo: string;
+}
+export interface ILPNavbarProps {
+
+}
+
+export interface ILPToastProps {
+  justifySelfEnd?: boolean;
+  textContent: string;
+}
